Add conditional compilation and radio station grid.
Conditional compilation uses VLC as the audio back-end when compiling
natively and uses wasm-bindgen web-sys as the back-end when compiling
webassembly. This was necessary because VLC is not compatible with
webassembly compilation, and web-sys is not compatible with native
compilation.

Also, a station grid was added for displaying radio station.